Updated test and seperated system properties
